move unimodal func
